adding calculBalance

+++ moneyManagementBackend/src/main/java/com/leopold/moneyManagemet/dto/LoginResponse.java
@@ -1,19 +1,19 @@
 package com.leopold.moneyManagemet.dto;
 
 public class LoginResponse {
-    private Long userId;
+    private int userId;
     private String message;
 
-    public LoginResponse(Long id, String message) {
+    public LoginResponse(int id, String message) {
         this.userId = id;
         this.message = message;
     }
 
-    public Long getUserId() {
+    public int getUserId() {
         return userId;
     }
 
-    public void setUserId(Long userId) {
+    public void setUserId(int userId) {
         this.userId = userId;
     }
