+++ moneyManagementBackend/src/main/java/com/leopold/moneyManagemet/models/SubscriberData.java
@@ -8,7 +8,7 @@ public class SubscriberData {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(name = "username")
     private String username;
@@ -63,11 +63,12 @@ public class SubscriberData {
         this.passwordConfirm = passwordConfirm;
     }
 
-    public Long getId() {
+
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 }
